requirements.txt
argon2-cffi==19.2.0
cffi==1.14.0
importlib-metadata==1.5.0
keyring==21.1.0
keyrings.cryptfile==1.3.4
pycparser==2.19
pycryptodome==3.9.6
pywin32-ctypes==0.2.0
six==1.14.0
zipp==3.0.0
